Facilitate catalog insertion with ingest.py
The IngestBase is meant to be inherited by the user, and provides a
methods for inserting observational catalogs (with potentially
multiple bandpasses per star observation) into Starplex.

IngestBase insures all relationships are made between
Bandpass, Catalog, CatalogStar and Observation tables. The
BandDefinitions class allows a bandpass to be specified, and return that
bandpass's database record if it already exists in the DB.