Connect returns engine and Session
Previously I attempted to store the engine and Session globally
in database.meta. However, it seems that these global variables can be
reset if the module if imported multiple times! Part of the problem is
that Starplex is intended to act as an application, but it is imported
by the primary application.

The solution is to make the base application responsible for holding
engine and Session after a connection. This Session() factory should be
passed to all code that needs to connect to the database.