RA/Dec points now encoded as PostGIS POINT columns
I'm using the Geography data type for PostGIS as this takes account of
spherical geometry, which the Geometry data type does not.

The API for the CatalogStar model has not changed; the RA/Dec columns
are combined internally.

A new module, starplex/database/tools.py is introduced that current
handles forming the POINT(ra, dec) strings for PostGIS. More misc
fuctions will be added there later.

Add a comment on PostGIS geometry vs geography and SRID