Implement SpatialJoin for joining with PostGis
The join algorithm implemented in the SpatialJoin class should be mostly
seen as a proof-of-concept, and is incredibly non-optimized.

Basic algorithm is to

1. For a given catalog to ingest, loop through all catalog stars,
   ordering from brightest to faintest (idea is that brightest stars should
   be matched first)
2. Do a PostGIS query to ask for all stars in the Star table near the
   catalog star.
3. Iterate through the matched stars and verify that they haven't
   already been matched to an object in this catalog. If not, add to the
   existing star
4. If the catalog cannot be matched, create a new entry in the Star
   table.

My hope was to at least combine queries 2 and 3 together in a single SQL
statement, but I could not get this to work. I tried to implement (3) as
a subquery within (2). However, I needed to use the ``in_`` method to
test if the catalog star's catalog is in the list of catalogs the star
is composed of. However, ``in_`` is not available.

Also implemented degree converted to meters (to deal with spatial
tolerance.)