Try passive_deletes=True to delete catalog stars
If I delete a Catalog row, I have postgres ON DELETE CASCADE set to
theoretically delete all catalog stars, and in turn, delete all
observations. With SQLAlchemy, to make this work I also need to set
`passive_deletes=True` in the relationship.

However, while this works for a Catalog deleting all referenced
CatalogStars, the referenced observations are not deleted, but just set
to have a foreign key of NULL. Thus the delete() method on Catalog does
extra deletes to clean up these orphans. It'd be great to figure out how
to make a two-level cascade delete work.