Add IntercalEdge table
IntercalEdge is essentially a directed graph containing zeropoint
offsets between connected fields of a single bandpass.

Each row is uniquely defined by the bandpass id, and the ids of the
'from' and 'to' catalogs of the overlap. The connections are directed to
indicate the order of the subtraction; delta is 'from - to'.