Solve network of field-to-field zeropoint corrections
We use scipy.optimize.basinhopping to get a global optimim. I've settled
on using Nelder-Mead optimization since it's reliable, but also slow.
The maximum number of function evaluations had to be boosted to allow NM
to converge. Much work can still be done in tuning step size and
annealing temperature.

Most crucial is converting the objective function to cython, though.